+++ modules/google-anyscale-iam/examples/variables.tf
@@ -39,3 +39,19 @@ variable "google_project_id" {
   description = "ID of the Project to put these resources in"
   type        = string
 }
+
+variable "existing_workload_identity_provider_name" {
+  description = <<-EOT
+    (Optional) Existing Workload Identity Provider Name.
+
+    The name of an existing Workload Identity Provider that you'd like to use. This can be in a different project.
+    If provided, this will skip creating a new Workload Identity Provider with the Anyscale IAM module.
+
+    ex:
+    ```
+    existing_workload_identity_provider_name = "projects/1234567890/locations/global/workloadIdentityPools/anyscale-existing-pool/providers/anyscale-existing-provider"
+    ```
+  EOT
+  type        = string
+  default     = null
+}